requirements.txt
playwright
playwright-stealth
beautifulsoup4
python-dotenv
gspread
google-auth
google-auth-oauthlib
